doc: Convert mainpage to Markdown
Convert to markdown. Our other GitHub files are in markdown
and doxygen natively supports it, so begin converting the
documentation entirely to markdown to minimize the number
of formats we have to comprehend.

Also rename it to index.md to match the other top level
files. Index and mainpage are interchangeable terms in
Doxygen.

Change-Id: I0f7274b47de6411f8d2d09d987fd1344e01a3986
Signed-off-by: Ben Walker <benjamin.walker@intel.com>