rte_virtio: remove unused static variable 'm'
The 'm' variable was only used in sizeof expressions; m was a struct
vhost_user_msg, and we already have one of those called 'msg' in the
function where it was needed.

A previous commit removed a RTE_USED() marking on m, which introduced a
warning on clang, since it can now see that m is unused.

Fixes: 73bfcede86d2 ("rte_virtio: remove rte_memzone from vq")

Change-Id: I63eb0d82c7c750b4181b6312bbda82ef3213c205
Signed-off-by: Daniel Verkamp <daniel.verkamp@intel.com>
Reviewed-on: https://review.gerrithub.io/389648
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Tested-by: SPDK Automated Test System <sys_sgsw@intel.com>