bdev/rbd: add semi-functional reset handling
Implement a poor approximation of reset handling for the RBD bdev so it
can pass the bdevio tests.

This should be replaced with a solution similar to bdev_aio, which waits
for all outstanding I/O to return before completing the reset.  However,
this should ideally be handled by the generic bdev layer rather than in
each individual module.

Change-Id: I1be2bf90875e23e05efdebb003e5412f3fc43793
Signed-off-by: Daniel Verkamp <daniel.verkamp@intel.com>
Reviewed-on: https://review.gerrithub.io/401273
Tested-by: SPDK Automated Test System <sys_sgsw@intel.com>
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>