autobuild.sh: fail build if Doxygen warnings found
We could set WARN_AS_ERROR = YES, but that will just
bail when the first error is found.  So instead just
fail the build in autobuild.sh and the submitter can
look at doxygen.log to see what went wrong.

Also change how doc make output is redirected to
doxygen.log, so that the commands show up in the
main build log instead of doxygen.log.  That
makes sure we can treat an empty doxygen.log file
as a successful test.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: Id3030d26e65c6c633b61855754e93b8b0b38b263
Reviewed-on: https://review.gerrithub.io/415873
Tested-by: SPDK Automated Test System <sys_sgsw@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-by: Daniel Verkamp <daniel.verkamp@intel.com>