go: empty Go package
The only purpose at the moment is to allow vendoring of the SPDK
source code into Go projects which manage dependencies with the "dep"
tool. That tool is designed for Go projects and as a sanity check
rejects any dependency which does not have at least one Go file, at
least at the moment (https://github.com/golang/dep/issues/1306).

Change-Id: I22458d0895729ad7d8ea53e7541e9089da1d448a
Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>
Reviewed-on: https://review.gerrithub.io/416831
Tested-by: SPDK Automated Test System <sys_sgsw@intel.com>
Reviewed-by: Daniel Verkamp <daniel.verkamp@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>