configure: fix building without isa-l submodule
The isa-l directory is generated automatically by
git on checkout, although it's empty. We check for
that directory existence in order to build ISA-L
and without `git submodule update --init` the build
just fails.

To fix, instead of checking the dir existence, we
check for the autogen.sh file inside that dir.

Change-Id: I7b62513afcd7e9f3953f825c41b265efe3484d57
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>
Reviewed-on: https://review.gerrithub.io/c/442523
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Piotr Pelpliński <piotr.pelplinski@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>