.travis.yml: only notify IRC for spdk/spdk repository
This avoids updates to forks of the main spdk repository
on GitHub from posting notifications to the #spdk channel.

I think this syntax should work, but I'm exactly sure
how to test outside of merging this patch, confirming
that spdk/spdk patches still result in notifications,
and creating an spdk fork and confirming there are no
notifications when it is updated.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: Ie5854d987b7e562abb73a36fd3ee10c43f740b5f

Reviewed-on: https://review.gerrithub.io/c/446317
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Reviewed-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>