examples/accel/perf: perf tool using the accel framework
Supports SW accel engine, CBDMA accel engine and eventually the DSA
accel engine through common API.  Basic functions only are included
here, subsequent patches will add more features as needed. For
example, the verify function is very crude right now.

Tool output was compared to the ioat perf tool w/CBDM and results
are very close although there a small penalty for using the
framework.

For CBDMA, start the app with --wait-for-rpc and run the ioat scan
RPC with desired whitelist. If there are not enough channels
(channels are not a concept in the framework) then SW accel engine
will be used for cores that could not get a channel.

Signed-off-by: paul luse <paul.e.luse@intel.com>
Change-Id: I9c5ab33e74db7bdbc5bfe5d7a08cc7ffaeab625b
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/883
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Reviewed-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>