autopackage:replace tarball and compile with debug build
Removed part of the script was only creating tarball from
sources and built from them.

Since SPDK is not distributed in this way usually,
this test does not add much value.

On the other hand this was the only palce where
release build (--disable-debug) was tested.

This patch replaces original tarballing,
with release build test as that was the only
part of the script adding value.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I6a054e15e6caf827653b671543c08d4a0986f353
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/1606
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-by: Karol Latecki <karol.latecki@intel.com>
Reviewed-by: Vitaliy Mysak <vitaliy.mysak@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>