test/ocf: Don't depend on gen_nvme.sh while detecting nvme ctrls
gen_nvme.sh requires that all the nvme devices are not bound to the
nvme driver which doesn't have to be the case when clear_nvme() is
called. To not depend on which driver is in use, simply iterate over
pci bus to find a proper device. This avoids additional setup.sh
ping-pong which is required to make use of gen_nvme.sh.

Additionally, upon exiting, in case mounts for given nvme device were
detected, keep the test environment reset to leave nvmes bound to the
nvme driver.

Change-Id: I428c5e2b0517553d267411a30316bde5d1ce8eef
Signed-off-by: Michal Berger <michalx.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/1843
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>