build: remove types suppressed from abidiff during 20.10
Some of these types had changes that we suppressed
during the 20.10 release cycle.  Now that 20.10 is
out, remove those suppressions since we will now
compare patches against the 20.10 release which
includes the affected changes.

Note that commit fce94287 filled out a new field
in spdk_nvme_ctrlr_data, so we will keep that
suppressed again for this release.  This is normal
and expected for spec-defined structures when we
define bits that were previously reserved.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I1a0f3b187330d443c4de7e013b4d66d1b8cb4e46
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/5234
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Reviewed-by: Aleksey Marchuk <alexeymar@mellanox.com>