scheduler_dynamic: simplify thread balancing logic
When picking a path to go down with a thread,
conditions unnecessarily piled up.

Instead do it either of two ways:
- move idle threads to main core
- find best core for active threads and move them there

There is no need to worry about cpumask of the thread,
since _find_optimal_core() will always return a core
within the cpumask.

If the found core is the same one as the current,
_move_thread() won't perform any action.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I0f4782766c15c86b5db0c970cfc9547058845b2a
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/8065
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>