vm_setup: add vfio-user to refspdk
refspdk is reference SPDK repository used for base in ABI tests.
When component is not present in the refspdk, but was added
in latest SPDK it is skipped in ABI tests.

Yet if the previous SPDK release contains such component already,
the ABI for it should be verified.

vfio-user flag was missing so this patch adds it.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I1e2c11c506deed06416aef7c0aab41d99834f92b
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/9127
Reviewed-by: Michal Berger <michalx.berger@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Reviewed-by: Monica Kenguva <monica.kenguva@intel.com>
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>