autobuild: group crypto patch apply for DPDK
Crypto patch for increasing maximum number of crypto
devs applies to all DPDK releases since 21.11.
Change the conditionals so that the patch is applied
to any new DPDK version.
Also change the patch location to "21.11+" directory
to indicate that patch is not exclusive for 21.11
version.

Change-Id: I341f88958c6bf6c4d77909a06bf7ce91200cb70e
Signed-off-by: Karol Latecki <karol.latecki@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/12159
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Reviewed-by: Michal Berger <michallinuxstuff@gmail.com>
Reviewed-by: Paul Luse <paul.e.luse@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@mellanox.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>