env_dpdk: remove -rpath-link from ENV_LINKER_ARGS
We already list the libraries with their explicit
pathnames, so the -rpath-link serves no purpose.

Our Makefile was actually specifying this option
without an = sign - i.e:

-Wl,-rpath-link /path/to/lib

On the submitter's system, this resulted in an error:

cc: Missing argument for -Wl,-rpath-link

I have no idea why no one has ever run into this
error, except for this one submitter.  But removing
the -rpath-link is the right thing to do here, since it
is not needed - so do that rather than adding the =
sign and continuing to figure out differences in
-Wl option processing on these different systems..

Fixes issue #2540.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I4f6176e55701a5dea5b10bba1ad621250cb5cb51
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/12984
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Dong Yi <dongx.yi@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>