spdk_top: allocate a copy of strings during thread display routine
The display_thread() function is the only one that needs to use a copy
of thread data, as threads might be moved between the cores during
SPDK application lifetime.
To avoid problems with using freed memory in the copied thread
structures, make sure the data is always present, by allocating a copy
of necessary strings. This will ensure that even after we have already
freed the globals holding data from RPC, the pointers used to display
the names and cpumasks will still have something to point to.

Fixes #2635

Change-Id: If9228ee199a774d208ddd6ff846532be3ef2012d
Signed-off-by: Krzysztof Karas <krzysztof.karas@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/14141
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Reviewed-by: Dong Yi <dongx.yi@intel.com>