configure: change the default ocf path
Change ocf default path to the rootdir relative one.
When --with-ocf flag is used in the configure,
being out of the spdk repository, the ocf path is
incorrectly set. Depends on the actual directory where
the configure is run.

The path should point to main spdk repo directory
regardless of where the configure is run from.

Signed-off-by: Kamil Godzwon <kamilx.godzwon@intel.com>
Change-Id: I6db3dd965ef9d119bb9e2f68ab3bd19c11b17a86
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/15022
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: wanghailiang <hailiangx.e.wang@intel.com>
Reviewed-by: Rafal Stefanowski <rafal.stefanowski@intel.com>
Reviewed-by: Karol Latecki <karol.latecki@intel.com>
Reviewed-by: Paul Luse <paul.e.luse@intel.com>
Reviewed-by: Krzysztof Karas <krzysztof.karas@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>