mk: replace [[ command with /dev/null redirection
make uses /bin/sh as its shell and some distributions set it to shells
that don't support [[ command (e.g. ubuntu and dash).  This causes the
check to fail with the following error:

/bin/sh: 1: [[: not found

So, replaced the [[ command with a redirection to /dev/null, which
should work on all shells.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Idcae094e8faa3da6faac16754bb0484bd91363d8
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/15852
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Mike Gerdts <mgerdts@nvidia.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Community-CI: Mellanox Build Bot